+++ Realm/RLMResults_Private.hpp
@@ -16,12 +16,36 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#import "RLMSyncPermissionResults.h"
+#import "RLMResults_Private.h"
 
 #import "results.hpp"
 
-@interface RLMSyncPermissionResults ()
+NS_ASSUME_NONNULL_BEGIN
 
+@interface RLMResults () {
+@protected
+    realm::Results _results;
+}
 - (instancetype)initWithResults:(realm::Results)results;
-
 @end
+
+NS_ASSUME_NONNULL_END
+
+// Utility functions
+
+[[gnu::noinline]]
+[[noreturn]]
+void RLMThrowResultsError(NSString * _Nullable aggregateMethod);
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wnullability-completeness"
+template<typename Function>
+static auto translateRLMResultsErrors(Function&& f, NSString *aggregateMethod=nil) {
+    try {
+        return f();
+    }
+    catch (...) {
+        RLMThrowResultsError(aggregateMethod);
+    }
+}
+#pragma clang diagnostic pop