+++ Realm/RLMSyncUser.mm
@@ -22,8 +22,9 @@
 #import "RLMNetworkClient.h"
 #import "RLMRealmConfiguration+Sync.h"
 #import "RLMRealmConfiguration_Private.hpp"
+#import "RLMResults_Private.hpp"
 #import "RLMSyncManager_Private.h"
-#import "RLMSyncPermissionResults_Private.hpp"
+#import "RLMSyncPermissionResults.h"
 #import "RLMSyncPermission_Private.hpp"
 #import "RLMSyncSession_Private.hpp"
 #import "RLMSyncSessionRefreshHandle.hpp"
@@ -54,7 +55,7 @@ NSError *translateExceptionPtrToError(std::exception_ptr ptr, bool get) {
     return error;
 }
 
-Permissions::PermissionResultsCallback RLMWrapPermissionResultsCallback(RLMPermissionResultsBlock callback) {
+std::function<void(Results, std::exception_ptr)> RLMWrapPermissionResultsCallback(RLMPermissionResultsBlock callback) {
     return [callback](Results results, std::exception_ptr ptr) {
         if (ptr) {
             NSError *error = translateExceptionPtrToError(std::move(ptr), true);