+++ Realm/Realm.h
@@ -35,7 +35,6 @@
 #import <Realm/RLMSyncPermission.h>
 #import <Realm/RLMSyncPermissionOffer.h>
 #import <Realm/RLMSyncPermissionOfferResponse.h>
-#import <Realm/RLMSyncPermissionResults.h>
 #import <Realm/RLMSyncSession.h>
 #import <Realm/RLMSyncUser.h>
 #import <Realm/RLMSyncUtil.h>